requirements.txt
fastapi==0.115.7
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
uvloop==0.21.0
openai-whisper==20240930
SpeechRecognition==3.14.2